Clean up and organize backend structure
- Remove duplicate models.py file (replaced by organized structure)
- Move config.py to db/ directory for better organization
- Update all imports to use new config location
- Clean up file structure with no duplicates
