--- apps/backend/src/models.py
@@ -1,66 +1,0 @@
-from sqlalchemy import Column, Integer, String, DateTime, Text, Boolean
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.sql import func
-from pydantic import BaseModel
-from datetime import datetime
-from typing import Optional
-
-Base = declarative_base()
-
-# Database Models
-class Image(Base):
-    __tablename__ = "images"
-    
-    id = Column(Integer, primary_key=True, index=True)
-    prompt = Column(Text, nullable=False)
-    generated_image_url = Column(String, nullable=True)
-    status = Column(String, default="pending")  # pending, processing, completed, failed
-    created_at = Column(DateTime(timezone=True), server_default=func.now())
-    updated_at = Column(DateTime(timezone=True), onupdate=func.now())
-    user_id = Column(Integer, nullable=True)  # For future user authentication
-    metadata_json = Column(Text, nullable=True)  # Store additional metadata as JSON
-
-class User(Base):
-    __tablename__ = "users"
-    
-    id = Column(Integer, primary_key=True, index=True)
-    email = Column(String, unique=True, index=True, nullable=False)
-    username = Column(String, unique=True, index=True, nullable=False)
-    hashed_password = Column(String, nullable=False)
-    is_active = Column(Boolean, default=True)
-    created_at = Column(DateTime(timezone=True), server_default=func.now())
-    updated_at = Column(DateTime(timezone=True), onupdate=func.now())
-
-# Pydantic Models for API
-class ImageCreate(BaseModel):
-    prompt: str
-    user_id: Optional[int] = None
-    metadata: Optional[dict] = None
-
-class ImageResponse(BaseModel):
-    id: int
-    prompt: str
-    generated_image_url: Optional[str]
-    status: str
-    created_at: datetime
-    updated_at: Optional[datetime]
-    user_id: Optional[int]
-    metadata_json: Optional[str]
-
-    class Config:
-        from_attributes = True
-
-class UserCreate(BaseModel):
-    email: str
-    username: str
-    password: str
-
-class UserResponse(BaseModel):
-    id: int
-    email: str
-    username: str
-    is_active: bool
-    created_at: datetime
-
-    class Config:
-        from_attributes = True
